Adding services, controller files of customer, manager class

+++ Bank-Management-System/src/main/java/com/bank/mgt/system/repository/AccountRepository.java
@@ -3,5 +3,6 @@ package com.bank.mgt.system.repository;
 import com.bank.mgt.system.entities.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountRepository extends JpaRepository<Account,Long> {
+public interface AccountRepository extends JpaRepository<Account,Integer> {
+
 }
