+++ Bank-Management-System/src/main/java/com/bank/mgt/system/repository/ManagerRepository.java
@@ -3,5 +3,5 @@ package com.bank.mgt.system.repository;
 import com.bank.mgt.system.entities.Manager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ManagerRepository extends JpaRepository<Manager,Long> {
+public interface ManagerRepository extends JpaRepository<Manager,Integer> {
 }
